requirements.txt
aenum==2.1.2
aiocontextvars==0.2.1
aiofiles==0.4.0
aioredis==1.2.0
aiosqlite==0.8.1
async-timeout==3.0.1
asyncpg==0.18.3
ciso8601==2.1.1
Click==7.0
contextvars==2.3
hiredis==1.0.0
httptools==0.0.11
immutables==0.6
multidict==4.5.2
PyPika==0.22.2
sanic==18.12.0
Sanic-Auth==0.2.0
sanic-session==0.5.0
tortoise-orm==0.11.2
ujson==1.35
websockets==6.0
Werkzeug==0.14.1
